requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.32.0
langgraph>=0.6.0
langchain>=0.2.0
langchain-core>=0.2.33
langchain-google-genai>=2.0.0
google-generativeai>=0.7.0
chromadb==0.4.24
python-dotenv==1.0.0
pydantic>=2.7.4
requests==2.31.0
sentence-transformers==2.3.1
lxml==4.9.4
PyPDF2==3.0.1
python-multipart==0.0.6
aiofiles==23.2.1
google-api-python-client==1.12.8